+++ doc/whats_new.rst
@@ -53,6 +53,8 @@ Changelog
 
 - Add option to unset a montage by passing `None` to :meth:`mne.io.Raw.set_montage` by `Clemens Brunner`_
 
+- Add sensor denoising via :func:`mne.preprocessing.oversampled_temporal_projection` by `Eric Larson`_
+
 - Add ``mne.io.pick.get_channel_types`` which returns all available channel types in MNE by `Clemens Brunner`_
 
 Bug